moves deps under api
