+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/api/containers_extension.py
@@ -11,13 +11,6 @@ from pydantic.main import BaseModel
 from servicelib.utils import logged_gather
 from simcore_sdk.node_ports_v2.port_utils import is_file_type
 
-from ..core.dependencies import (
-    get_application,
-    get_mounted_volumes,
-    get_rabbitmq,
-    get_settings,
-    get_shared_store,
-)
 from ..core.docker_logs import start_log_fetching, stop_log_fetching
 from ..core.docker_utils import docker_client
 from ..core.rabbitmq import RabbitMQ
@@ -28,6 +21,13 @@ from ..models.schemas.ports import PortTypeName
 from ..modules import directory_watcher, nodeports
 from ..modules.data_manager import pull_path_if_exists, upload_path_if_exists
 from ..modules.mounted_fs import MountedVolumes
+from ._dependencies import (
+    get_application,
+    get_mounted_volumes,
+    get_rabbitmq,
+    get_settings,
+    get_shared_store,
+)
 from .containers import send_message
 
 logger = logging.getLogger(__name__)
